Move VideoCallSharedPerferencesPersistence to video.calling package.
